+++ croud/organizations/commands.py
@@ -102,3 +102,41 @@ def organizations_delete(args: Namespace) -> None:
 
     if errors is None and args.org_id == CONFIG.organization:
         CONFIG.set_organization_id(CONFIG.name, None)
+
+
+def org_file_uploads_list(args: Namespace) -> None:
+    client = Client.from_args(args)
+    data, errors = client.get(f"/api/v2/organizations/{args.org_id}/file-uploads/")
+    print_response(
+        data=data,
+        errors=errors,
+        keys=["id", "name", "upload_url"],
+        output_fmt=get_output_format(args),
+    )
+
+
+def org_file_uploads_create(args: Namespace) -> None:
+    client = Client.from_args(args)
+    data, errors = client.post(f"/api/v2/organizations/{args.org_id}/file-uploads/")
+    # TODO: Handle file upload
+    print_response(
+        data=data,
+        errors=errors,
+        success_message="File upload created. Please use the URL provided to upload "
+        "your file.",
+        keys=["id", "name", "upload_url"],
+        output_fmt=get_output_format(args),
+    )
+
+
+def org_file_uploads_delete(args: Namespace) -> None:
+    client = Client.from_args(args)
+    data, errors = client.delete(
+        f"/api/v2/organizations/{args.org_id}/file-uploads/{args.file_id}"
+    )
+    print_response(
+        data=data,
+        errors=errors,
+        success_message="File upload deleted.",
+        output_fmt=get_output_format(args),
+    )